cloud_functions/update_listing_cache/requirements.txt
google-cloud-datastore==1.9.0
requests==2.22.0
pydantic==1.6.1
beautifulsoup4==4.8.0
lxml==4.4.1
